Refactoring the PostgreSQL Virtual Network Rule
